+++ common/security/src/main/java/com/hotels/road/security/LdapSecurityConfiguration.java
@@ -19,9 +19,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisSentinelConfiguration;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.ldap.core.support.BaseLdapPathContextSource;
 import org.springframework.ldap.core.support.LdapContextSource;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
@@ -33,7 +30,7 @@ import org.springframework.security.ldap.search.FilterBasedLdapUserSearch;
 import org.springframework.security.ldap.search.LdapUserSearch;
 import org.springframework.security.ldap.userdetails.DefaultLdapAuthoritiesPopulator;
 import org.springframework.security.ldap.userdetails.LdapAuthoritiesPopulator;
-import org.springframework.session.data.redis.config.annotation.web.http.EnableRedisHttpSession;
+import org.springframework.session.hazelcast.config.annotation.web.http.EnableHazelcastHttpSession;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
@@ -45,18 +42,8 @@ import avro.shaded.com.google.common.collect.ImmutableList;
 @Profile("ldap")
 @Configuration
 @EnableGlobalMethodSecurity(prePostEnabled = true)
-@EnableRedisHttpSession
+@EnableHazelcastHttpSession
 public class LdapSecurityConfiguration {
-  @Bean
-  public RedisConnectionFactory redisConnectionFactory(
-      @Value("${redis.master}") String redisMaster,
-      @Value("${redis.sentinelHost}") String redisSentinelHost,
-      @Value("${redis.sentinelPort}") int redisSentinelPort) {
-    RedisSentinelConfiguration sentinelConfig = new RedisSentinelConfiguration().master(redisMaster).sentinel(
-        redisSentinelHost, redisSentinelPort);
-    return new LettuceConnectionFactory(sentinelConfig);
-  }
-
   @Bean
   public BaseLdapPathContextSource contextSource(
       @Value("${ldap.url}") String ldapUrl,
@@ -92,7 +79,8 @@ public class LdapSecurityConfiguration {
       @Value("${ldap.group.searchSubtree}") boolean searchSubtree,
       @Value("${ldap.group.ignorePartialResultException}") boolean ignorePartialResultException,
       @Value("${ldap.group.searchFilter}") String groupSearchFilter) {
-    DefaultLdapAuthoritiesPopulator authoritiesPopulator = new DefaultLdapAuthoritiesPopulator(contextSource,
+    DefaultLdapAuthoritiesPopulator authoritiesPopulator = new DefaultLdapAuthoritiesPopulator(
+        contextSource,
         groupSearchBase);
     authoritiesPopulator.setSearchSubtree(searchSubtree);
     authoritiesPopulator.setIgnorePartialResultException(ignorePartialResultException);
@@ -138,8 +126,7 @@ public class LdapSecurityConfiguration {
   }
 
   @Bean
-  public WebSecurityConfigurerAdapter webSecurityConfigurerAdapter(
-      LdapAuthenticationProvider authenticationProvider) {
+  public WebSecurityConfigurerAdapter webSecurityConfigurerAdapter(LdapAuthenticationProvider authenticationProvider) {
     return new RoadWebSecurityConfigurerAdapter() {
       @Override
       public void configure(AuthenticationManagerBuilder auth) throws Exception {